server_django/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
connection==2021.7.20
Django==4.1.7
djangorestframework==3.14.0
Flask==2.2.3
Flask-Cors==3.0.10
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
mysqlclient==2.1.1
numpy==1.24.2
pandas==1.5.3
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sqlparse==0.4.3
threadpoolctl==3.1.0
tzdata==2022.7
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
